% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrplotting.R
\name{add.dist.to}
\alias{add.dist.to}
\title{add new covariate consisting of distances to points that have covariate==distance.to}
\usage{
add.dist.to(mask, covariate, distance.to, dname = NULL, overwrite = FALSE)
}
\arguments{
\item{mask}{is object of class "mask".}

\item{covariate}{is name of one of the elements of covariates(mask)}

\item{distance.to}{is the target value of covariate; distances to mask points with this value are calculated.}

\item{dname}{is the name you want to give to the new covariate}
}

