% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcdplotting.R
\name{add_lcpath}
\alias{add_lcpath}
\title{Add leas cost path to plot}
\usage{
add_lcpath(mask, cname, cfun = "mean", directed = FALSE, symm = TRUE,
  conductance = FALSE, ...)
}
\arguments{
\item{mask}{\code{secr} mask object. Must have covariate called 'noneuc' containing cost}

\item{cname}{Name of variable to use in cost calculation}

\item{cfun}{Cost function name. Defauls to "mean"}

\item{directed}{If TRUE, use directed graph for transition between cells, else undirected}

\item{symm}{If TRUE, cost is same in both directions}

\item{conductance}{If TRUE, cfun must calculate conductance (the inverse of cost), not cost.}
}
\description{
Calculates and plots the least-cost path between two points according 
 to the function \code{cfun}, using the mask covariate \code{cname}.
 
 Requires packages raster, gdistance, igraph
}
\details{
Invisibly returns an object of class \code{SpatialLines} defining the least cost path.
}
