% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrplotting.r
\name{addRaster2Mask2}
\alias{addRaster2Mask2}
\title{Adds covariate to mask from raster object}
\usage{
addRaster2Mask2(ras, mask, maskCRS, covname = NULL)
}
\arguments{
\item{ras}{is an object of class `RasterLayer'}

\item{mask}{is an object of class `mask'}

\item{maskCRS}{is an object of class "CRS" giving the coordinate reference 
system of the mask coordinates}

\item{covname}{optional argument specifying what the covariate added to the
mask is to be called.}
}
\description{
Adds covariate to mask from raster object
}
