% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{beta2natural}
\alias{beta2natural}
\title{Transforms secr object parameters from 'beta' to natural scale}
\usage{
beta2natural(fit)
}
\arguments{
\item{fit}{An \code{secr} fitted object.}
}
\description{
Transforms 'beta' parameters, SEs and confidence bounds onto natural scale.
}
\details{
This function adapts the unexported \code{secr} functions \code{secr:::untransform} and
 \code{secr:::se.untransform} to operate directly on an object of class \code{secr}.
}
