% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{Calc.AICc}
\alias{Calc.AICc}
\title{AICc}
\usage{
Calc.AICc(mod.objs_, decimals_ = 2, n.inds_)
}
\arguments{
\item{mod.objs_}{A list of the model objects.}

\item{decimals_}{The number of decimals to round the value to.}

\item{n.inds_}{The number of detected individuals in the data set.}
}
\value{
Returns the AICc values.
}
\description{
Calculates the AICc for a given model or set of models.
}
