% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{Calc.ER.Variance}
\alias{Calc.ER.Variance}
\title{Calculate the variance for the time-dependent expected encounter rate.}
\usage{
Calc.ER.Variance(par.ests, num.spline.pars, vcov.mat, type = "lambda.t",
  method = "delta", Breps = 1000, alpha = 0.05, resolution = 100,
  dist = NULL, pred.mesh = NULL)
}
\arguments{
\item{par.ests}{The estimated parameters from the chosen model (excluding Density). The order of the parameters (on the link scale)
needs to be: lambda0, sigma, and then the spline parameters.}

\item{num.spline.pars}{The number of estimated spline parameters.}

\item{vcov.mat}{The estimated variance-covariance matrix, from inverting the hessian. The matrix should exclude sigma if the Delta 
method is chosen for the lambda0(t) function.}

\item{type}{Specifies the model parameterisation as either "lambda.t" or "sigma.t".}

\item{method}{This argument specifies the method that is used to approximate the variance, the Delta method (`delta') or a 
parametric bootstrap approach (`boot').}

\item{Breps}{Sets the number of bootstrap replications.}

\item{alpha}{Sets the desired level of confidence for the intervals that are returned.}

\item{resolution}{The temporal resolution of the points.}

\item{dist}{A single value for distance. If this is NULL the function assumes the time-dependent part of the function.}

\item{pred.mesh}{Allows the function to be plotted for any vector of points i.e. not just for a single cycle.}
}
\value{
Returns a matrix of lower and upper limits (for a given alpha value) for the estimated ER or sigma(t) / lambda0(t) at a 
vector of time points. If the bootstrap approach is used, the estimated ERs from each bootstrap replication are also returned whereas the
if the Delta method is chosen the estimated standard errors (on the ER scale) are also returned.
}
\description{
Calculates the variance for the estimated time-dependent ER that is used to plot confidence intervals. The 
time component of the ER is a simple linear combination of spline parameters, whereas the full encounter rate function is 
a non-linear function of the spline and the detection function parameters. Both the Delta method and a parametric bootstrapping 
approach is used to estimate the variance of the full encounter rate function.
}
