% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSCRfunctions.R
\name{calc.p.Pis}
\alias{calc.p.Pis}
\title{Calculates p..(s) and P_i(s) for SCR likelihood calculation.}
\usage{
calc.p.Pis(capthist, mesh, pars, dist)
}
\arguments{
\item{capthist}{Capture history object, or library \code{secr} class 'capthist.}

\item{mesh}{Integration mesh, of library \code{secr} class 'mask': an Mx2 matrix.}

\item{pars}{Vector of paramters, named "g0", "sigma" and "D".}

\item{dist}{KxM matrix of distances between the K detectors in traps(capthist) and the M mesh points.}
}
\value{
A list with these elements: 
\itemize{
 \item{p..s}{Probability of detection by any detector, for activity centres at each mesh point 
 (vecotr of length M)}
 \item{Pi.s}{For each of n detected individuals, the rows of this matrix contain the probability of 
 obtaining the observed capture history, for all M possible activity centre locations (mesh points) 
 (nxM matrix).}
 }
}
\description{
Calculates p..(s) (a vector of length n) and P_i(s) an nxM matrix of probabilities for 
SCR likelihood calculation.
}

