% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{Calc.pos.Haz}
\alias{Calc.pos.Haz}
\title{Calcluates the position in the beta vector for all parameters.}
\usage{
Calc.pos.Haz(D.mod_ = c(1), D.type_ = c("Cons"), lambda0.mod_ = c(1),
  b.effect_ = c("n"), sigma.mod_ = c(1), haz.type_ = "Dep",
  haz.mod_ = c(1), haz.k_ = c(4), conduct.mod_ = NULL)
}
\arguments{
\item{D.mod_}{A numeric vector that specifies what density parameters are required for the different sessions. A different 
number specifies different parameters e.g. if there are 2 sessions: c(1,1) specifies the same density parameters for the two 
sessions whereas c(1,2) specifies that the 2nd session needs its own density parameters. The length of the vector must equal 
the number of sessions.}

\item{D.type_}{A string vector that specifies the density model for each session. Options include constant, exponential 
and quadratic density ( c("Cons","Exp", "Quad"). The length of the vector must equal the number of sessions in the data.}

\item{lambda0.mod_}{A numeric vector that specifies what intercept parameters are required for the different sessions. A different 
number specifies a different parameter e.g. if there are 2 sessions: c(1,1) specifies the same lambda0 parameter for the two 
sessions whereas c(1,2) specifies that the 2nd session needs its own lambda0 parameter. The length of the vector must equal 
the number of sessions.}

\item{b.effect_}{A string vector  that specifies if there is a behavioural effect on the intercept of the encounter rate / 
detection function. Options include c("n","b") where "b" indicates the effect. The length of the vector must equal the number 
of sessions in the data.}

\item{sigma.mod_}{A numeric vector that specifies what range parameters are required for the different sessions. A different 
number specifies a different parameter e.g. if there are 2 sessions: c(1,1) specifies the same sigma parameter for the two 
sessions whereas c(1,2) specifies that the 2nd session needs its own sigma parameter. The length of the vector must equal 
the number of sessions.}

\item{haz.type_}{The type of detection hazard to use in the model. Choices c("IndL", "IndU", "Dep") include the two 
independent hazard parameterisations and the dependent hazard parameterisation.}

\item{haz.mod_}{A numeric vector that specifies what spline parameters are required for the different sessions. A different 
number specifies different parameters e.g. if there are 2 sessions: c(1,1) specifies the same spline parameters for the two 
sessions whereas c(1,2) specifies that the 2nd session needs its own spline parameters. The length of the vector must equal 
the number of sessions.}

\item{haz.k_}{The values here match what is passed to mgcv and so do not correspond with the number of spline parameters that
are estimated. The length of the vector must equal the number of sessions in the data. A value of 1 specifies a constant hazard 
and then no spline parameters are involved.}

\item{conduct.mod_}{A numeric vector that specifies whether the same conductance parameter is used for the different sessions. 
A different number specifies that different parameters are estimated. The length of the vector must equal the number of sessions.}
}
\value{
Returns a list with the positions in the beta vector for density and encounter rate / detection function parameters.
}
\description{
The function calcluates the position in the beta vector for density, encounter rate / detection function, 
and spline parameters that depends on the type of model that is specified. The function also accommodates the sharing 
of parameters across different sessions if desired.
}
