% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{calcEn}
\alias{calcEn}
\title{Calculates expected number of captures.}
\usage{
calcEn(fit)
}
\arguments{
\item{fit}{An \code{secr} fitted model. (This must have an element \code{$mask}.)}
}
\description{
Calculates the expected number of captures from an \code{secr} object.
}
\examples{
require(secr)
calcEn(secrdemo.0)

}
