% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{CDF.gam.cyc}
\alias{CDF.gam.cyc}
\title{Construct the CDF for capture times.}
\usage{
CDF.gam.cyc(
  ind.haz_,
  study.vec_,
  hazt.vec.tot_,
  detector.type_ = "Prox",
  haz.type_ = "Ind"
)
}
\arguments{
\item{ind.haz_}{The individual and detector-specific detection hazard(s). For a dependent hazard, the argument will be a vector 
of the actual encounter rates whereas for an independent hazard it will be a scalar representing lambda_d. For SC traps the
argument will be a vector of values across trpas that gets summed for an overall hazard.}

\item{study.vec_}{A vector of times that correspond to the time period one wants to simulate captures for.}

\item{hazt.vec.tot_}{The h(t) component of the encounter rate function i.e. the detection hazard at different cycle times.
Only relevant for independent hazards.}

\item{detector.type_}{Specifies the type of detector. Defaults to using a "proximity" or passive detector but can also simulate
from an array of single-catch traps.}

\item{haz.type_}{The type of detection hazard to use in the simulation. The choice is between an independent hazard ("Ind) 
or a dependent hazard ("Dep).}
}
\value{
Returns a matrix of dimension number traps * time steps that represents F(t) or the prob of being caught before time t.
}
\description{
Constructs the Cumulative Distribution Function (CDF) for capture times conditional on a frequency or 
count of captures. The conditional nature opf the CDF means it must end at 1 and hence a time will always be generated.
Called by the simctsCH function.
}
