% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcdplotting.R
\name{closest_coords}
\alias{closest_coords}
\title{Finds closest coordinates on raster to two points}
\usage{
closest_coords(from, to, rastermask)
}
\arguments{
\item{from}{pair of coordinates (x,y) from which to start}

\item{to}{pair of coordinates (x,y) to get to}

\item{rastermask}{Raster object (typically created from mask by something like 
rastermask = raster(mask,"noneuc"))}
}
\value{
Returns the coordinates of the closest point on the raster, as a matrix with two columns (x,y), 
named s1 and s2, with first row corresponding to 'from' coordinates, and second row corresponding to 'to' 
coordinates.
}
\description{
Uses function over() from package sp to overlay points on raster and return closest raster coordinates
}
