% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condDensity.r
\name{condDensity}
\alias{condDensity}
\title{Conditional value of fiited density smooth}
\usage{
condDensity(fit, covariate, otherlevels = NULL, mask = NULL,
  se = TRUE, cl = TRUE, nx = 200)
}
\arguments{
\item{fit}{An object of class \code{secr}.}

\item{covariate}{A characted giving the name of the selected covariate (which must be in \code{names(covariates(mask))}; see below).}

\item{otherlevels}{A named vector with the selected level/value of each of the other covariates used in \code{fit} 
(defaults to their mean). (NOTE: Use of this argument has not been properly tested - things might go wrong!)}

\item{mask}{An object of class \code{mask}, which contains the covariates 'covariate' and those for 'otherlevels'.}

\item{se}{TRUE if you want standard errors of density calculated and returned.}

\item{nx}{Number of values of 'covariate' to use.}

\item{lcl}{TRUE if you want the lower 95\% bound of density calculated and returned.}

\item{se}{TRUE if you want the upper 95\% bound of density calculated and returned.}
}
\value{
A list comprising:
\itemize{
 \item{preds}{ A data frame with these columns:
 \itemize{
 \item{x}{ Values of the chosen covariate at each of \code{nx} points spanning its range.}
 \item{Dhat}{ Estimated density at each value of \code{x}.}
 \item{se}{ Standard error of \code{Dhat}.}
 \item{lcl}{ Lower 95\% confidence bound for density at each value of \code{x}.}
 \item{ucl}{ Upper 95\% confidence bound for density at each value of \code{x}.}
 }}
 \item{detcovs}{ Values of the chosen covariate at each detector (in order of detectors in \code{traps(fit)})}
 \item{condvals}{ Named vector with values of each covariate being conditioned on.}
 \item{xvar}{ Name of the selected covariate.}
 }
}
\description{
Calculates the estimated density over the range of values of one spatial covariate, at
fixed values of all other spatial covariates.
}
\examples{
\dontrun{
library(scrmlebook)
data("leopard-sim") # simulated Kruger Park data
# Fit overly complex density model - just for illustration:
fit = secr.fit(kruger.capt,model=list(D~x+s(y,k=3)+s(habitat.cov,k=3)),kruger.mask)
smest = condDensity(fit,covariate="y") # calculate conditional density
Dscale = 100^2 # scale up to number per sq km
# Plot conditional density in Northing dimension
plot(smest$preds$x,smest$preds$Dhat*Dscale,ylim=range(0,na.omit(smest$preds[,-1]))*Dscale,
     xlab="Northing",ylab=expression(hat(D)),type="n")
lines(smest$preds$x,smest$preds$Dhat*Dscale)
lines(smest$preds$x,smest$preds$lcl*Dscale,lty=2)
lines(smest$preds$x,smest$preds$ucl*Dscale,lty=2)
# add locations of camera traps
points(smest$detcovs,rep(0,length(smest$detcovs)),pch="|",cex=0.75,col="black")
# Plot conditional density in dist.to.water dimension
smest = condDensity(fit,covariate="habitat.cov") # calculate conditional density
plot(smest$preds$x,smest$preds$Dhat*Dscale,ylim=range(0,na.omit(smest$preds[,-1]))*Dscale,
     xlab="habitat.cov",ylab=expression(hat(D)),type="n")
lines(smest$preds$x,smest$preds$Dhat*Dscale)
lines(smest$preds$x,smest$preds$lcl*Dscale,lty=2)
lines(smest$preds$x,smest$preds$ucl*Dscale,lty=2)
# add locations of camera traps
points(smest$detcovs,rep(0,length(smest$detcovs)),pch="|",cex=0.75,col="black")
}
}
\seealso{
\code{\link{condDensityplot}}
}
