% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrplotting.r
\name{condDensityplot}
\alias{condDensityplot}
\title{Plot conditional value of fiited density smooth}
\usage{
condDensityplot(fit, covariate, otherlevels = NULL, mask = NULL,
  se = TRUE, cl = TRUE, nx = 200, Dscale = 100^2,
  xlab = covariate, ylab = expression(hat(D)), main = "",
  Dlcol = "black", Dlty = 1, Dlwd = 1, CIlcol = "black",
  CIlty = 2, CIlwd = 1, tkcol = "black", tkcex = 0.75,
  tkpch = "|")
}
\arguments{
\item{fit}{An object of class \code{secr}.}

\item{covariate}{A characted giving the name of the selected covariate (which must be in \code{names(covariates(mask))}; see below).}

\item{otherlevels}{A named vector with the selected level/value of each of the other covariates used in \code{fit} 
(defaults to their mean). (NOTE: Use of this argument has not been properly tested - things might go wrong!)}

\item{mask}{An object of class \code{mask}, which contains the covariates 'covariate' and those for 'otherlevels'.}

\item{se}{TRUE if you want standard errors of density calculated and returned.}

\item{nx}{Number of values of 'covariate' to use.}

\item{Dscale}{A multiplier to scale density: the defaule Dscale=100^2 gives density in number per square km.}

\item{xlab}{Label for x-axis. Defaults to the \code{covariate} argument value.}

\item{lcl}{TRUE if you want the lower 95\% bound of density calculated and returned.}

\item{se}{TRUE if you want the upper 95\% bound of density calculated and returned.}
}
\value{
Invisibly returns a list comprising:
\itemize{
 \item{preds}{ A data frame with these columns:
 \itemize{
 \item{x}{ Values of the chosen covariate at each of \code{nx} points spanning its range.}
 \item{Dhat}{ Estimated density at each value of \code{x}.}
 \item{se}{ Standard error of \code{Dhat}.}
 \item{lcl}{ Lower 95\% confidence bound for density at each value of \code{x}.}
 \item{ucl}{ Upper 95\% confidence bound for density at each value of \code{x}.}
 }}
 \item{detcovs}{ Values of the chosen covariate at each detector (in order of detectors in \code{traps(fit)})}
 \item{condvals}{ Named vector with values of each covariate being conditioned on.}
 \item{xvar}{ Name of the selected covariate.}
 }
}
\description{
Calculates and plots the estimated density over the range of values of one spatial covariate, at
fixed values of all other spatial covariates. This function is pretty inflexible but also pretty simple. If you want
to do a more flexible plot, just take the few lines of code inside the function and cahnge some things.
}
\examples{
\dontrun{
library(scrmlebook)
data("leopard-sim") # simulated Kruger Park data
# Fit overly complex density model - just for illustration:
fit = secr.fit(kruger.capt,model=list(D~x+s(y,k=3)+s(habitat.cov,k=3)),kruger.mask)
smest = condDensityplot(fit,covariate="y") 
# Plot conditional density in dist.to.water dimension
smest = condDensityplot(fit,covariate="habitat.cov") 
}
}
\seealso{
\code{\link{condDensity}}
}
