% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{create.GAM}
\alias{create.GAM}
\title{Create gam object using mgcv.}
\usage{
create.GAM(time.vec_, haz.vec_, GAM.k_ = 4, pred.mesh_ = NULL)
}
\arguments{
\item{time.vec_}{Vector of x values for the detection hazard cycle. Usually this vector is fairly short
since it requires a matching number of haz.vec_ values in order to specify the desired shape.}

\item{haz.vec_}{Vector of values for the hazard corresponding to the different hazard cycle times.
The values supplied will determine the shape of the detection hazard.}

\item{GAM.k_}{The specified degrees of freedom for the regression spline. This number corresponds to what gets passed to mgcv.
Note that one degree of freedom is lost for a cyclic cubic spline and hence the actual number of parameters that get estimated is one less than the supplied value.}

\item{pred.mesh_}{Specifies the time vector used for prediction / plotting. This vector is different 
to the time.vec_ argument in order to produce smooth plots.}
}
\value{
Returns the gam object, the spline coefficients, the fitted glm object, and the basis 
functions corresponding to both the vector of prediction times and the vector of cycle times.
}
\description{
Creates a GAM object from supplied x and y values that specify a detection hazard.
Fits a cyclic cubic regression spline with the specified d.o.f.
}
