% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{CT.SCR}
\alias{CT.SCR}
\title{Constructs the Continuous-Time (CT) SCR likelihood.}
\usage{
CT.SCR(beta_, data_, trap.objs_, mask.objs_, dists_ = NULL,
  D.type_ = c("Cons"), D.mod_ = c(1), lambda0.mod_ = c(1),
  sigma.mod_ = NULL, lambda0.b_ = c("n"), haz.mod_ = c(1),
  haz.type_ = "IndL", haz.k_ = c(6), conduct.mod_ = NULL,
  fixed.b0_ = 0, usage_ = NULL, endpoint_, setup.hour_ = 0,
  occ.interval_ = 24, cycle.vec_ = seq(0, 24, by = 1),
  mesh.size_ = 100, trace_ = FALSE)
}
\arguments{
\item{beta_}{A vector of parameter values. These contain the starting values for the optimisation when passed to nlm.}

\item{data_}{The data frame. The capture times variable needs to be called "Caught", the detector column "Detector", and 
the identifier column "id".}

\item{trap.objs_}{The secr trap object. This argument needs to be a list when a model is fitted to more than one session.}

\item{mask.objs_}{The secr mask object. This argument needs to be a list when a model is fitted to more than one session.}

\item{dists_}{A matrix of distances between the detectors and the mask points. If this is NULL it will be created by the function.}

\item{D.type_}{A string vector that specifies the density model to be fitted for each session. Options include constant, exponential 
and quadratic density ( c("Cons","Exp", "Quad"). The length of the vector must equal the number of sessions in the data.}

\item{D.mod_}{A numeric vector that specifies whether the same density model and parameter(s) is used for the different sessions.
A different number specifies that different parameters are estimated. The length of the vector must equal the number of sessions in 
the data.}

\item{lambda0.mod_}{A numeric vector that specifies whether the same intercept parameter for the encounter rate / detection function
is used for the different sessions. A different number specifies that different parameters are estimated. The length of the vector 
must equal the number of sessions in the data.}

\item{sigma.mod_}{A numeric vector that specifies whether the same range parameter for the encounter rate / detection function
is used for the different sessions. A different number specifies that different parameters are estimated. The length of the vector 
must equal the number of sessions in the data.}

\item{lambda0.b_}{Specifies if an overall behavioural effect on the intercept parameter for the encounter rate / detection function
is required. Not developed yet for trap-specific effects or for the range parameter.}

\item{haz.mod_}{A numeric vector that specifies whether the same spline parameters for the encounter rate / detection function
is used for the different sessions. A different number specifies that different parameters are estimated. The length of the vector 
must equal the number of sessions in the data.}

\item{haz.type_}{The type of detection hazard to use in the model. Choices c("IndL", "IndU", "Dep") include the two 
independent hazard parameterisations and the dependent hazard parameterisation.}

\item{haz.k_}{A numeric vector that specifies the df for the regression spline for the different sessions. A different number 
specifies different df's for different sessions. A constant haz is specified by giving a value of 1,and can be used for either
the linked or unlinked independent hazard. The length of the vector must equal the number of sessions in the data.}

\item{conduct.mod_}{A numeric vector that specifies whether the same conductance parameter is used for the different sessions. 
A different number specifies that different parameters are estimated. The length of the vector must equal the number of sessions in the data.}

\item{fixed.b0_}{A scalar for the b0 spline parameter. Although it is relevant for the independent hazards, it is redundant
due to the fact that lambda_d scales the estimated encounter rate.}

\item{usage_}{A list that supplies information about trap usage i.e. between what time periods each detector was NOT operating.}

\item{endpoint_}{The endpoint of the study that determines the study duration. The length of the vector must equal the number of sessions in the data.}

\item{setup.hour_}{The starting point of the study. Values of zero are assumed unless alternative values are supplied, in which case
the length of the vector must equal the number of sessions in the data.
Can be used if the survey started after time 0 i.e. if it does not start at the start of a hazard cycle.}

\item{occ.interval_}{The duration of an occasion. Primarily used when integrating the hazard function over one occasion.}

\item{cycle.vec_}{A vector of times that correspond to a single hazard cycle.}

\item{mesh.size_}{Determines the resolution of the time vector. This argument determines the vector length from the seq() command.}
}
\value{
Returns the negative log likelihood value. The chosen optimiser used to find parameter estimates is actually a minimiser
and minimising the negative log likelihood is equivalent to maximising the log likelihood.
}
\description{
Constructs the CT likelihood that is required to fit these models using maximum likelihood and a half normal form 
for the encounter rate / detection function.
}
