% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamethod.R
\name{DeltaMethod}
\alias{DeltaMethod}
\title{Numeric Delta Method approximation for the variance-covariance matrix}
\usage{
DeltaMethod(par, fct, vcov, delta = 0.01, ...)
}
\arguments{
\item{par}{vector of parameter values at which estimates should be constructed.}

\item{fct}{function that constructs estimates from parameters \code{par}.}

\item{vcov}{variance-covariance matrix of the parameters.}

\item{delta}{proportional change in parameters used to numerically estimate 
first derivative with central-difference formula.}

\item{...}{any additional arguments needed by \code{fct}.}
}
\value{
Returns a list with values
\itemize{
 \item{variance}{ Estimated variance-covariance matrix of estimates derived
 by \code{fct}.}
 \item{partial}{ Matrix (or vector) of partial derivatives of\code{fct} with
 respect to the parameters  \code{par}.}
}
}
\description{
Computes delta method variance-covariance matrix of results of any generic 
 function fct that computes a vector of estimates as a function of a set of 
 estimated parameters par.
}
\details{
This function is copied vebatim from package \link{mrds}.

The delta method (aka propagation of errors is based on Taylor series 
approximation). It uses 
the first derivative of fct with respect to par which is computed in this 
function numerically using the central-difference formula. It also uses the 
variance-covariance matrix of the estimated parameters which is derived in 
estimating the parameters and is an input argument.

The first argument of fct should be par which is a vector of parameter 
estimates. It should return a single value (or vector) of estimate(s). The 
remaining arguments of fct if any can be passed to fct by including them at 
the end of the call to DeltaMethod as name=value pairs.
}
\examples{

par = c(-4.77, 2.64)
vcv = matrix(c(0.067,-0.054, -0.054, 0.068),nrow=2,byrow=TRUE)
fn = function(par) invlogit(sum(par))
DeltaMethod(par=beta.g0.d, fct=invlogit, vcov=vcv)

}
