% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{Devdf}
\alias{Devdf}
\title{Calculates Deviance/df statistic.}
\usage{
Devdf(fit, ...)
}
\arguments{
\item{fit}{An \code{secr} fitted model. (This must have an element \code{$mask}.)}

\item{...}{Unused argument, here to make use in \code{gof.sim} easy (as some functions
passed to \code{gof.sim} have other arguments, and this dummy argument allows the same
syntax to be used for them and this.)}
}
\description{
Calculates the scaled deviance divided by the model degrees of freedom.
}
\examples{
require(secr)
Devdf(secrdemo.0)

}
