% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSCRfunctions.r
\name{distances}
\alias{distances}
\title{Calculates distances between two sets of points in the plane}
\usage{
distances(X, Y)
}
\arguments{
\item{X}{is an A by 2 matrix.}

\item{Y}{is a B by 2 matrix.}
}
\details{
Calculates distances between points in X and Y, whic are 
2-column matrices of coordinates (x- and y-coordinates in the two columns).
Returns a matrix with dimensions A=dim(X)[1] and B=dim(Y)[1], containing the distances.
}
\references{
This function written by Murray Efford
}

