% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erdetfun.R
\name{erdetfun}
\alias{erdetfun}
\title{Calculates encounter or detection function values and CIs}
\usage{
erdetfun(x, newdata = NULL, dists = 0:200, rgr = FALSE, ci = FALSE,
  ci.method = "secr", alpha = 0.05, what = "p", B = 999)
}
\arguments{
\item{x}{an object of class \code{secr}.}

\item{newdata}{a data frame with a column containing appropriate covariate
values for each of the covariates in \code{x} (both density model and 
detection or encounter model covariates, even though denstiy covariates are 
not used for detection function calculation).}

\item{dists}{distances at which to calculate the detection function.}

\item{rgr}{TRUE if you want distance multiplied by detection function value
to be returned (as per point transects, for example).}

\item{ci}{If TRUE, confidence intervals are calculated and returned.}

\item{ci.method}{If 'secr' then the method of package \code{secr} (i.e. a kind
of Delta Method) is used to calculate confidence intervals, else if 'bootstrap' 
nonparametric bootstrap, resampling from a multivariate normal with mean 
equal to the estimated model parameters and variance-covariance equal to 
their estimated variance-covariance matrix, is used.}

\item{alpha}{Significance level for confidence intervals}

\item{what}{If 'p' the detection function is returned; if 'er' the encounter 
function is returned.}

\item{B}{number of bootstrap samples to take}
}
\description{
Calculates an encounter function or a detection function at specified distances, 
 given an \code{secr} object, a data frame containing covariates used by the 
 object, and the distances.
 Returns a data frame of x (distance), y (detection function) and (optionally) 
 lower and upper confidence interval values, calculated either using the 
 method of the package \code{secr}, or by parameteric bootstrap..
}
