% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamethod.R
\name{ermodel.g0}
\alias{ermodel.g0}
\title{Delta Method approximation for g(0) from model with lambda0}
\usage{
ermodel.g0(model)
}
\arguments{
\item{fit}{An object of class \code{secr}.}
}
\value{
Returns a list with values
\itemize{
 \item{g0}{ Estimated g0 parameter.}
 \item{se.g0}{ Std error of g0 estimate.}
}
}
\description{
Computes delta method  (first-order Taylor series approximation) for 
 g0 parameter and its variance from a model parameterised in terms of lambda0.
}
