% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{Freeman_Tukey}
\alias{Freeman_Tukey}
\title{Calculates Freeman-Tukey GoF statistic.}
\usage{
Freeman_Tukey(object, En.capt = NULL)
}
\arguments{
\item{object}{An object of class \code{secr} fitted model (containing an element \code{$mask}), 
or an object of class \code{capthist}.}

\item{En.capt}{The expected number of captures in total.}
}
\description{
Calculates the Freeman-Tukey GoF statistic. If \code{object} is of class \code{secr}, then 
\code{En.capt} is ignored and the expected number of captures is calculated from \code{object} 
and the statistic is calculated from  \code{object$capthist}. If If \code{object} is of class 
\code{capthist} then \code{En.capt} must contain the expected number of captures and the 
statistic is calculated from this and \code{capthist}.
}
\examples{
require(secr)
Freeman_Tukey(secrdemo.0)

}
