% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamethod.R
\name{g0model.lambda0}
\alias{g0model.lambda0}
\title{Delta Method approximation for lambda0 from model with g0}
\usage{
g0model.lambda0(model)
}
\arguments{
\item{fit}{An object of class \code{secr}.}
}
\value{
Returns a list with values
\itemize{
 \item{lambda0}{ Estimated lambda0 parameter.}
 \item{lambda0}{ Std error of lambda0 estimate.}
}
}
\description{
Computes delta method  (first-order Taylor series approximation) for 
 lambda0 parameter and its variance from a model parameterised in terms of g0.
}
