% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_simtest}
\alias{gof_simtest}
\title{Simulate sample of a GoF test statistic.}
\usage{
gof_simtest(
  object,
  statfn = Devdf,
  nsim = 99,
  newfit = FALSE,
  plot = TRUE,
  nbins = 20,
  seed = NULL
)
}
\arguments{
\item{object}{An \code{secr} fitted model. (This must have an element \code{$mask}.)}

\item{statfn}{A function that calculates a  GoF test statistic from a fitted model or 
\code{capthist} object.}

\item{nsim}{The number of simulations to do.}

\item{newfit}{If TRUE, refits model to every simulated dataset and recalculates expected numbers.}

\item{plot}{If TRUE, produces a histogram of simulated statistics with observed statistic marked.}

\item{nbins}{Number of histogram bins in plot}

\item{seed}{Random number seed (setting it gives repeatability of results).}
}
\description{
Repeatedly simulates a population and capture history using the parameters of a 
fitted \code{secr} model as truth, fits a model of the same sort to the data,
and calculates the test statistic (specified via\code{statfn}) for each new
fitted model.

Only works with full likelhood models (those that include a model for density), and 
does not work for all of them. In particular, it works for models without
individual effects, except that it does work for a model with sex as a 
partially-observed individual effect in which both sexes have a common density.

Has only been tested with single-occasion surveys.
}
\examples{
require(secr)
gofdbn1 = gof_simtest(secrdemo.0,nsim=99,seed=1)
gofdbn1$pvalue

}
