% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrplotting.r
\name{implotcovariate}
\alias{implotcovariate}
\title{Plots perspective image of mask covariate value}
\usage{
implotcovariate(mask, covariate, ...)
}
\arguments{
\item{mask}{is an object of class \code{mask}}

\item{covariate}{is a character variable with the name of one of the covariates in mask (the one to plot)}

\item{...}{other arguments to be passed to \code{persp}}
}
\description{
Plots perspective image of mask covariate value
}
