% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{intervals.outage}
\alias{intervals.outage}
\title{Create outage intervals.}
\usage{
\method{intervals}{outage}(trap.name, usage_)
}
\arguments{
\item{trap.name}{The trap name that is passed to this function by sapply. The trap names are extracted as rownames of the trap object.}

\item{usage_}{A list that supplies information about trap usage i.e. between what time periods each detector was NOT operating.}
}
\value{
Returns the intervals that are of class "Intervals" created by the intervals package.
}
\description{
creates intervals for the outage times that are supplied with the usage information using the intervals R package.
The function is used with sapply and the trap names.
}
