% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcdplotting.R
\name{lcpathplot}
\alias{lcpathplot}
\title{Least-cost path plot}
\usage{
lcpathplot(mask, transitionFunction, type = "noneuc", n = 512,
  background = "d2.river", linecol = "white", directions = 16,
  symm = TRUE, directed = FALSE, lwd = 1, ...)
}
\arguments{
\item{mask}{A mask with covariate `noneuc` (obtained using \code{predictDsurface})}

\item{transitionFunction}{A user-defined function for calculating conductance between points 
(see function \code{transition} of the \code{gdistance} package).
'userdfn1' in the \code{secr} vignette 'secr-noneuclidean.pdf', for example.)}

\item{n}{number of least-cost paths to plot}

\item{linecol}{color of least cost path line}

\item{directions}{parameter of function \code{transition}}

\item{symm}{parameter of function \code{transition}}

\item{directed}{parameter of function \code{transition}}

\item{...}{arguments to function \code{plotcovariate}, which is used to plot the image.}

\item{covariate}{The name (a character) of a covariate to plot as an image plot
(typically because 'noneuc' depends on it).}
}
\description{
Plots 
 Requires packages secr, raster, gdistance, fields, igraph
}
\examples{
lcpathplot(Dhat.Nemegt,"stdBC",userdfun=function(x) 1/x,n=5,contour=FALSE,col=c("orange","brown"),key=FALSE, linecol="white",xaxt="n",yaxt="n",xlab="",ylab="",bty="n",asp=1)
}
