% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{LL.captimes}
\alias{LL.captimes}
\title{Calculate the log likelihood value for captures.}
\usage{
LL.captimes(x_, dists_ = NULL, log.lambda.d_ = NULL,
  log.lambda.b.d_ = NULL, lambda0_, lambda0.b_, const.haz_ = F,
  gam.obj_, spline.coeff_)
}
\arguments{
\item{x_}{The function is used with apply() on each individual's capture history. The first argument x_ is a
row of data from the capture history data frame.}

\item{dists_}{A matrix of distances between detectors and mask points. Used when a dependent hazard is specified.}

\item{log.lambda.d_}{A matrix with the log encounter rates for each detector corresponding to each mask point, i.e. the rates 
exclude the time component. Only supplied for independent hazards, if it is NULL a dependent hazard is implied.}

\item{log.lambda.b.d_}{Another log lambda matrix corresponding to distances but using a different parameter estimate after 
first capture. Only relevant for models that have a behavioural effect specified.}

\item{lambda0_}{The intercept parameter for the dependent hazard encounter rate function.}

\item{lambda0.b_}{A second intercept parameter for the dependent hazard encounter rate function that is applicable after first
capture. Only relevant for models that have a behavioural effect specified.}

\item{const.haz_}{Indicates a hazard that does not cvary through time.}

\item{gam.obj_}{GAM obj used to find the detection hazard for the capture times.}

\item{spline.coeff_}{A vector of coefficients to be used in the regression spline.}
}
\value{
Returns a vector for each capture time of dimension equal to the number of mask points.
}
\description{
calculates the contribution to the log likelihood term for each capture history. Called by the 
CT.SCR function and used with the apply() function.
}
