% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{LL.Exposure}
\alias{LL.Exposure}
\title{Calculate the extent of exposure to each detector.}
\usage{
LL.Exposure(
  x_,
  haz.type_ = "IndL",
  lambda.d_ = NULL,
  full.exp_,
  const.haz_ = F,
  dists_ = NULL,
  traps.fail.file_,
  lambda0_ = NULL,
  study.time_,
  gam.obj_,
  spline.coeff_
)
}
\arguments{
\item{x_}{The function is used with apply() and the trap names and so the first argument x_ is the 
trap name that gets passed by apply().}

\item{haz.type_}{The type of detection hazard to use in the model. Choices c("IndL", "IndU", "Dep") include the two 
independent hazard parameterisations and the dependent hazard parameterisation.}

\item{lambda.d_}{If this is Null then it implies that a dependent hazard is used}

\item{full.exp_}{The total exposure (integrated hazard) corresponding to a full hazard cycle for all detectors.}

\item{const.haz_}{Indicates if a constant hazard is used.}

\item{dists_}{A matrix of distances between detectors and mask points.}

\item{traps.fail.file_}{A list of trap usage information.}

\item{lambda0_}{Value for the intercept of the encounter rate function. Only needed when a dependent hazard is specified.}

\item{study.time_}{A vector corresponding to the study duration.}

\item{gam.obj_}{GAM obj used to predict the detection hazard for the outage periods.}

\item{spline.coeff_}{A vector of coefficients to be used in the regression spline.}
}
\value{
Returns a vector of total exposure
}
\description{
Called by the CT.SCR function and used to calculate the total exposure for each trap. Usage information will
obviously reduce this exposure for traps not operating for the full survey duration.
}
