% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{LL.survival.ind}
\alias{LL.survival.ind}
\title{Calcluates the exposure to capture for individuals.}
\usage{
LL.survival.ind(haz.type_, trap.obj_, first.cap_,
  usage.intervals_ = NULL, lambda.d_ = NULL, lambda.b.d_ = NULL,
  lambda0_ = NULL, lambda0.b_ = NULL, const.haz_ = F,
  dists_ = NULL, gam.obj_ = NULL, spline.coeff_ = NULL, study.time_,
  setup.hour_ = 0, endpoint_, mask.pts_)
}
\arguments{
\item{haz.type_}{The type of detection hazard to use in the model. Choices c("IndL", "IndU", "Dep") include the two 
independent hazard parameterisations and the dependent hazard parameterisation.}

\item{trap.obj_}{An secr trap object.}

\item{first.cap_}{The time of first capture for a specific individual.}

\item{usage.intervals_}{Intervals of outages that are of class "Intervals". Required to subtract out periods of exposure that
correspond to detectors being out of action.}

\item{lambda.d_}{A matrix with the encounter rates for each detector corresponding to each mask point, i.e. the rates 
exclude the time component. Only supplied for independent hazards, if it is NULL a dependent hazard is implied.}

\item{lambda.b.d_}{Another lambda matrix corresponding to distances but using a different parameter estimate after 
first capture. Only relevant for models that have a behavioural effect specified.}

\item{lambda0_}{The intercept parameter for the dependent hazard encounter rate function.}

\item{lambda0.b_}{A second intercept parameter for the dependent hazard encounter rate function that is applicable after first
capture. Only relevant for models that have a behavioural effect specified.}

\item{const.haz_}{Indicates if a constant hazard is used.}

\item{dists_}{A matrix of distances between detectors and mask points. Used when a dependent hazard is specified.}

\item{gam.obj_}{GAM obj used to find the detection hazard for the capture times.}

\item{spline.coeff_}{A vector of coefficients to be used in the regression spline.}

\item{study.time_}{A vector corresponding to the study duration.}

\item{setup.hour_}{Can be used if the survey started after time 0 i.e. if it does not start at the start of a hazard cycle.}

\item{endpoint_}{The endpoint of the study, this vzlue determines the study duration.}

\item{mask.pts_}{Gives the number of rows or cells in the mask object.}
}
\value{
Returns a vector of total exposure.
}
\description{
Calcluates the exposure to capture for individuals and is required for models with a behavioural effect whereby
each individual's exposure to detection changes depending on when they were first caught. The function also uses the usage 
to reduce exposure related to trap outages.
}
