% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{makeDtable}
\alias{makeDtable}
\title{Summarises density estimates from secr objects}
\usage{
makeDtable(object, ..., criterion = c("AICc", "AIC"), Dscale = 1)

makeDtable(object, ..., criterion = c("AICc", "AIC"), Dscale = 1)
}
\arguments{
\item{...}{Other object of class '\code{secrlist}' or '\code{secr}'.}

\item{criterion}{Character variable being 'AIC' or 'AICc'.}

\item{Dscale}{Numeric scaling factor for density (which is multiplied by this factor).}

\item{allargs}{Object of class '\code{secrlist}' or '\code{secr}' from package \code{secr}.}
}
\description{
Makes a data frame summarising density estimates, ordered by AIC or AICc.

Makes a data frame summarising density estimates, ordered by AIC or AICc.
}
\details{
Lists the following in each column of the data frame, for each model, with models arranged in increasing order of AIC or AICc
(whichever was specified in the call). Density estimate, SE, lower- and upper 95\% confidence intervals are multiplied by 
argument \code{Dscale}
\itemize{
 \item{Dhat:}{ Density estimate.}
 \item{se:}{ Standard error of density estimate.}
 \item{lcl:}{ Lower 95\% confidence bound for density.}
 \item{ucl:}{ Upper 95\% confidence bound for density.}
 \item{cv:}{ Coefficient of variation of density estimate.}
 \item{npar:}{ Number of parameters in the model.}
 \item{logLik:}{ Log-Likelihood}
 \item{AIC:}{ AIC or AICc (whichever was specified in the call).}
 \item{dAIC:}{ Delta AIC or Delta AICc (whichever was specified in the call).}
 \item{AICwt:}{ AIC weight or AICc weight (whichever was specified in the call).}
 }
 Requires function \code{AIC} from package \code{secr}.

Lists the following in each column of the data frame, for each model, with models arranged in increasing order of AIC or AICc
(whichever was specified in the call). Density estimate, SE, lower- and upper 95\% confidence intervals are multiplied by 
argument \code{Dscale}
\itemize{
 \item{Dhat:}{ Density estimate.}
 \item{se:}{ Standard error of density estimate.}
 \item{lcl:}{ Lower 95\% confidence bound for density.}
 \item{ucl:}{ Upper 95\% confidence bound for density.}
 \item{cv:}{ Coefficient of variation of density estimate.}
 \item{npar:}{ Number of parameters in the model.}
 \item{logLik:}{ Log-Likelihood}
 \item{AIC:}{ AIC or AICc (whichever was specified in the call).}
 \item{dAIC:}{ Delta AIC or Delta AICc (whichever was specified in the call).}
 \item{AICwt:}{ AIC weight or AICc weight (whichever was specified in the call).}
 }
 Requires function \code{AIC} from package \code{secr}.
}
