% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{noneuc.distances}
\alias{noneuc.distances}
\title{Calcluates the least cost distances.}
\usage{
noneuc.distances(traps.obj_, mask.obj_, transitionFn_ = 1,
  directions_ = 16, conduct.par_ = 0)
}
\arguments{
\item{traps.obj_}{A traps object that is used to get least cost distances from every trap to every mask cell.}

\item{mask.obj_}{A mask object that needs to have one numeric covariate attached to it.}

\item{transitionFn_}{Specifies how the conductance between two cells is calculated. Defaults to 1 for the arithmetic mean.
2 specifies the geometric mean. Used in gdistance's transition function.}

\item{directions_}{Specifies how many different transitions are possible from each cell. Used in gdistance's transition function.}

\item{conduct.par_}{Supplies the conductance parameter that is used with the spatial copvariate to calculate the noneuc 
linear predictor.}
}
\value{
Returns a matrix containing least cost distances between all cells.
}
\description{
The function calcluates the least cost distance matrix for a given conductance parameter and spatial covariate.
Assumes one covariate attached to the mask object. exp(Beta0) is sigma and beta0 has been excluded. Requires the gdistance
package.
}
