% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcdplotting.R
\name{noneucdists}
\alias{noneucdists}
\title{Least-cost distance summary}
\usage{
noneucdists(masknu, userfn, covariate = NULL)
}
\arguments{
\item{masknu}{A mask with covariate `noneuc` (obtained using \code{predictDsurface})}

\item{userfn}{A user-defined function for calculating least-cost distances (see function 
'userdfn1' in the \code{secr} vignette 'secr-noneuclidean.pdf', for example.)}

\item{covariate}{The name (a character) of a covariate to summarise with the distances 
(typically because 'noneuc' depends on it).}
}
\description{
Tabulates least-cost distance between smallest and largest `noneuc` values on mask
 Requires packages secr, raster, gdistance, fields, igraph
}
