% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erdetfun.R
\name{noplot.secr}
\alias{noplot.secr}
\title{Calculates detection function values}
\usage{
noplot.secr(x, newdata = NULL, rgr = FALSE, limits = FALSE,
  alpha = 0.05, xval = 0:200)
}
\arguments{
\item{x}{an object of class \code{secr}.}

\item{newdata}{a data frame with a column containing appropriate covariate
values for each of the covariates in \code{x} (both density model and 
detection or encounter model covariates, even though denstiy covariates are 
not used for detection function calculation).}

\item{rgr}{TRUE if you want distance multiplied by detection function value
to be returned (as per point transects, for example).}

\item{alpha}{Significance level for confidence intervals}

\item{xval}{distances at which to calculate the detection function.}
}
\description{
Calculates a detection function at specified distances, given an \code{secr}
 object, a data frame containing covariates used by the object, and the 
 distances
 This is a hacked version of the \link{secr} function \code{plot.secr} - 
 hence the name '\code{noplot.secr}', The hack is so as not to plot, and to 
 return a data frame of x (distance), y (detection function) and (optionally) 
 lower and upper confidence interval values.
}
