% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrplotting.r
\name{plot.Dsurface}
\alias{plot.Dsurface}
\title{Plots image (and optionally contours) of density from Dsurface object}
\usage{
\method{plot}{Dsurface}(Dsurface, covariate = "D.0", scale = 1,
  contour = TRUE, addpoly = TRUE, ...)
}
\arguments{
\item{Dsurface}{is an object of class `Dsurface'}

\item{covariate}{The covariate to plot (density by default)}

\item{scale}{Factor to multiply surface by for plotting}

\item{contour}{is a logical, TRUE if want contour plots on image}

\item{addpoly}{If TRUE, adds bounding polygon}

\item{...}{other arguments to be passed to \code{prep4image}}
}
\description{
Plots image (and optionally contours) of density from Dsurface object
}
