% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrplotting.r
\name{plot.er.fit}
\alias{plot.er.fit}
\title{Encounter rate fit plot}
\usage{
plot.er.fit(scrfit, dmax, binwidth, lwd = 2, xlab = "x", ylab = "y",
  hcol = "red", hlineonly = FALSE, houtline = FALSE, hfill = TRUE)
}
\arguments{
\item{scrfit}{Object of class  '\code{secr}' from package \code{secr}. The function only 
works with traps of class '\code{count}'.}

\item{dmax}{Maximum distance to plot.}

\item{binwidth}{width of each distance interval to plot.}

\item{lwd}{Line width for encounter rate function.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{hcol}{Colour of histogram.}

\item{hlineonly}{If TRUE, uses \code{\link{lines}} to draw histogram lines on current plot; 
else uses \code{\link{plot}} to draw lines on new plot.}

\item{houtline}{If TRUE, draws only the outline (profile) of the histogram; else draws each 
complete bar.}

\item{hfill}{TRUE, fills bards with \code{hcol}, else leaves them empty.}
}
\description{
Plots mean encounter rate by distance interval, with estimated encounter rate function
 overlaid. NB: Distances calculated from estimated activity centre mode. Don't over-interpret
 the plot - the activity centre locations may have very high uncertainty attached to them, 
 and hence the distances may be very uncertain!
 
 The function only works with traps of class '\code{count}'.
 
 Uses function '\code{histline}'.
}

