% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrplotting.r
\name{plotcovariate}
\alias{plotcovariate}
\title{Plots image (and optionally contours) of mask covariate value}
\usage{
plotcovariate(mask, covariate, contour = TRUE, key = TRUE, ...)
}
\arguments{
\item{mask}{is an object of class `mask'}

\item{covariate}{is a character variable with the name of one of the covariates in mask (the one to plot)}

\item{contour}{is a logical, TRUE if want contour plots on image}

\item{...}{other arguments to be passed to \code{prep4image}}
}

