% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrplotting.r
\name{plotDline}
\alias{plotDline}
\title{Plots density as function of a covariate}
\usage{
plotDline(
  fit,
  cov,
  mask = NULL,
  per.km = TRUE,
  add.traps = TRUE,
  sess = NULL,
  logD = FALSE,
  Dlim = NULL
)
}
\arguments{
\item{fit}{An object of class "secr".}

\item{cov}{A character variable of length 1 giving name of covariate to use 
for the x-axis. This must be the name of a covariate on the mask}

\item{mask}{If not NULL, is object of class "mask". If NULL, takes from fit.}

\item{per.km}{If TRUE, converts density to number per square km, else per hectare.}

\item{add.traps}{TRUE if you want the locations of traps to appear as ticks on the x-axis.}

\item{sess}{Number of session to use (if NULL, first session is used)}

\item{logD}{If TRUE, log(Density+1) is plotted instead of Density}

\item{Dlim}{vector containing min and max Density to plot (y-axis is restircted to this range)}
}
\description{
Plots density as function of a covariate
}
