% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrplotting.r
\name{plotNpc}
\alias{plotNpc}
\title{Plot top proportion of density}
\usage{
plotNpc(Dmask, pc, boundary = NULL, binary = FALSE, ...)
}
\arguments{
\item{Dmask}{is object of class "Dsurface".}

\item{pc}{is top percentage of the abundance to plot. Must be betwen 0 and 100. 
For example, if p=25, the distribution of the highest density regions that accounts for 
25% of the total abundance is plotted.}

\item{boundary}{is a SpatialPolygonsDataFrame defining the boundary of the sruvey region.}

\item{binar}{is a variable, which if true, causes the top pc to be assigned a single 
density value of 1, rather than its actual density.}
}
\description{
Plot top proportion of density
}
