% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{pred.Y}
\alias{pred.Y}
\title{Predict detection hazard}
\usage{
pred.Y(time.vec_, gam.obj_, fit.obj_ = NULL, spline.coeff_ = NULL)
}
\arguments{
\item{time.vec_}{Vector of x values for the detection hazard cycle.}

\item{gam.obj_}{A gam object created using package mgcv. Required in order to extract the basis functions.}

\item{fit.obj_}{A fitted glm object. The spline coefficients are extracted from the glm object when simulating.}

\item{spline.coeff_}{A vector of coefficients to be used in the regression spline.}
}
\value{
Returns the hazard values.
}
\description{
Uses a gam object to predict encounter rate / hazard of detection that correspond to the x values supplied.
}
