% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{predetectfn}
\alias{predetectfn}
\title{Calculates encounter or detection function.}
\usage{
predetectfn(
  dists,
  detectfn,
  pars,
  effort = rep(1, nrow(dists)),
  hazard = TRUE,
  rgr = FALSE
)
}
\arguments{
\item{dists}{a matrix of distances.}

\item{detectfn}{\code{secr} detection function type (number or character).}

\item{pars}{list like that returned by the \code{secr} function \code{detectpar}.}

\item{effort}{Trapping effort (multiplies hazard). Must be a vector of length
\code{nrow(dist)}.}

\item{hazard}{If TRUE, returns encounter function, else detection function.}

\item{rgr}{If TRUE multiplies encounter or detection function by distance.}
}
\description{
Calculates encounter function or detection function, scaled by distance or not.

Calculates encounter function or detection function at given \code{dists}.
}
\examples{
detectors <- make.grid (nx = 6, ny = 8, detector = "multi")
detections <- sim.capthist (detectors, popn = list(D = 10, buffer = 100), 
                            detectpar = list(g0 = 0.2, sigma = 25))
## fit & print null (constant parameter) model
model <- secr.fit (detections)
dists = matrix(c(0,10,20,20,10,0),nrow=2,byrow=2)
pars = detectpar(model)
detectfn = model$detectfn
predetectfn(dists,detectfn,pars)
predetectfn(dists,detectfn,pars,effort=c(1,2)))
predetectfn(dists,detectfn,pars,effort=c(1,2)),hazard=FALSE)
predetectfn(dists,detectfn,pars,hazard=FALSE)

}
