% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrplotting.R
\name{prep4image}
\alias{prep4image}
\title{Prepares data frame for plotting with image/contour/persp.}
\usage{
prep4image(data, plot = TRUE, contour = TRUE, key = TRUE, ...)
}
\arguments{
\item{data}{a data frame with columns x and y being Cartesian coordinates, 
and z being the values of some variable at each coordinate.}

\item{plot}{if \code{TRUE} then an image plot will be drawn using 
\code{\link{image.plot}}}

\item{contour}{if \code{TRUE} then contours will be added (only used when 
\code{plot=TRUE})}

\item{key}{logical for whether or not to include key when \code{plot = TRUE} (\code{\link{image.plot}} is used when \code{key = TRUE}, \code{\link{image}} is used when \code{key = FALSE})}

\item{...}{other arguments to pass to \code{\link{image}} or \code{\link{image.plot}} (only used 
when \code{plot=TRUE})}
}
\description{
From an input data frame with columns x, y and z, this function 
  creates a list with elements x, y and z in a format suitable for passing to
  functions \code{\link{image}}, \code{\link{contour}} or 
  \code{\link{persp}}. The coordinates in \code{data} are assumed to come
  from a 2D grid of points.
}
\details{
Sorts z on values of x first, then y, then creates a matrix of 
  z-values from this. Returns a list with elements x (unique values of x, in 
  increasing order), y (unique values of y, in increasing order) and z 
  (matrix of z-values in appropriate order for image/contour/persp). 
  
  If the original z is a factor variabele, the z returned is a matrix of integers 
  between 1 and length(levels(z)) and the output list has an attributes called 
  ``facnames'' that is a character vector containing the levels as factor 
  variables, with z=1 corresponding to the first name, z=2 to the second, etc.
}

