% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{prob.detect}
\alias{prob.detect}
\title{Calculates the probability of detection from a chosen activity centre, for a given time point and for a given duration.}
\usage{
prob.detect(pixel_, mask.obj_, beta1_, duration_, time_, lambda0_,
  spline.coeff_, sigma_ = NULL)
}
\arguments{
\item{pixel_}{The two dimensional coordinate that corresponds to the pixel of interest.}

\item{mask.obj_}{A mask object that needs to have one numeric covariate attached to it.}

\item{beta1_}{The conductance parameter.}

\item{duration_}{The length of time that the detection probability is caluclated for.}

\item{time_}{A supplied time in the 24 hour cycle.}

\item{lambda0_}{A value for the base encounter rate at a distance of zero/}

\item{spline.coeff_}{A vector of spline coefficients that determine the shape of the temporal encounter rate function.}

\item{sigma_}{A value for the sigma parameter in the encounter rate function. This needs to be supplied for the lambda(t) 
function. If it is NULL the sigma(t) form will be assumed.}
}
\value{
Returns a vector of capture probabilities associated with each pixel in the mask.
}
\description{
The function calculates the detection probability from a chosen pixel for a given time point and for a 
given duration. The probability depends on the mask of the study area, the covariate attached to the mask, and a given 
conductance parameter.
}
