% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{remove.duplicates}
\alias{remove.duplicates}
\title{Remove duplicate times.}
\usage{
remove.duplicates(data.frame_, n.inds_)
}
\arguments{
\item{data.frame_}{The capture history data.}

\item{n.inds_}{The number of unique individuals in the capture history data.}
}
\value{
Returns the capture histories without the duplicate times..
}
\description{
Gets rid of any duplicate times that may have been simulated, for the same individual at the same detector. 
This function is called by sim.ctsCH. Note that it requires variables called "Caught" and "id" in the data frame.
}
