% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{reorder}
\alias{reorder}
\title{Reorders a vector of id's.}
\usage{
reorder(id.vec)
}
\arguments{
\item{id.vec}{A vector of individual ids.}
}
\value{
Returns a vector of reordered ids.
}
\description{
This function is called by sim.ctsCH and reorders the id numbers in the simulated data so that the numbers run
from 1 to n.
}
