% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSCRfunctions.R
\name{scr.negloglik}
\alias{scr.negloglik}
\title{SCR negative log-likelihood}
\usage{
scr.negloglik(pars, capthist, mesh, dist = NULL)
}
\arguments{
\item{pars}{Vector of paramters, named "g0", "sigma" and "D".}

\item{capthist}{Capture history object, or library \code{secr} class 'capthist'.}

\item{mesh}{Integration mesh, of library \code{secr} class 'mask': an Mx2 matrix.}

\item{dist}{KxM matrix of distances between the K detectors in traps(capthist) and the M mesh points.}
}
\value{
Value of negative log-likelihood
}
\description{
Calculates the log-likelihood for an SCR model. Currently only multi-catch traps and 
count detectors with a single session implemented, and no detection function or density model covariates implemented.
}

