\name{scrmlebook-package}
\alias{scrmlebook-package}
\alias{scrmlebook}
\docType{package}
\title{
Utility package for book SCR by maximum likelihood, with R
}
\description{
Functions and datasets to go with the book 'Spatial Capture-Recapture by Maximum Likelihood, with R'.
}
\details{
\tabular{ll}{
Package: \tab scrmlebook\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-10-09\cr
License: \tab GNU General Public License Version 2 or later\cr
}
Functions and datasets to go with the book 'Spatial Capture-Recapture by Maximum Likelihood, with R'.
}
\author{
David Borchers et al.

Maintainer: David Borchers <dlb@st-andrews.ac.uk>
}
\references{
Borchers, D.L., Stevenson, B.C., Rexstad, E., Howe, E.J. and Distiller, G. Spatial Capture-Recapture by Maximum Likelihood, with R.
}
\keyword{Spatial Capture-Recapure}
\examples{
To be added
}