% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrmlebook.R
\docType{package}
\name{scrmlebook}
\alias{scrmlebook}
\alias{scrmlebook-package}
\title{Utility package for book SCR by maximum likelihood, with R}
\description{
Functions and datasets to go with the book 'Spatial Capture-Recapture by Maximum Likelihood, with R'.
}
\examples{
## To be added.
}
\author{
David Borchers et al.
}
\references{
Borchers, D. L., Stevenskon, B. C., Rexstad, E., Howe, E. J., and Distiller, G. Spatial Capture-Recapture by Maximum Likelihood, with R
}
\keyword{Spatial}
\keyword{capture-recapture}

