% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{shadedCI}
\alias{shadedCI}
\title{Draws a shaded confidence region for a line}
\usage{
shadedCI(xyci, col = adjustcolor("gray", alpha.f = 0.4), border = NA,
  ...)
}
\arguments{
\item{xyci}{a data frame or matrix with first element being x-coordinates, 
second being the corresponding lower (or upper) confidence bound values, 
third being the corresponding upper (or lower) confidence bound values.}

\item{col}{an argument of \code{polygon} specifying shading colour.}

\item{border}{an argument of \code{polygon} specifying border colour}

\item{...}{other arguments to \code{polygon}.}
}
\description{
Adds a shaded polygon covering the region between lower and upper confidence
 bounds, to an existing plot.
}
