% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{sim.ctsCH}
\alias{sim.ctsCH}
\title{Simulate CT data.}
\usage{
sim.ctsCH(traps.obj_, mask.obj_ = NULL, buffer_ = NULL,
  haz.type_ = "IndL", detector.type_ = "Prox", gam.obj_ = NULL,
  fit.obj_ = NULL, cycle.hrs_ = c(0, 24), mesh.size_ = 1000,
  dens.surf_, g0_ = NULL, lambda0_ = NULL, sigma_ = NULL, endpoint_,
  setup.hour_ = 0, occ.interval_ = 24)
}
\arguments{
\item{traps.obj_}{An secr traps object.}

\item{mask.obj_}{An secr mask object. If not supplied make.mask will be used to create the mask.}

\item{buffer_}{Buffer radius added to traps.obj_ to create the mask (if not supplied).}

\item{haz.type_}{The type of detection hazard to use in the simulation. Choices c("IndL", "IndU", "Dep") include the two 
independent hazard parameterisations and the dependent hazard parameterisation.}

\item{gam.obj_}{GAM obj to simulate capture times from according to a time varying spline hazard.}

\item{fit.obj_}{A fitted glm object. The spline coefficients are extracted from the object when simulating.}

\item{cycle.hrs_}{The start and end of a single detection hazard cycle, defaults to c(0,24) to represent a daily cycle.}

\item{mesh.size_}{Determines the resolution of the time vector. This argument determines the vector length from the seq() command.}

\item{dens.surf_}{The density to use in the simulation. If a single number then the data will be simulated with a constant density.
The argument can also be a matrix of values of the same dimension as the mask that is supplied.}

\item{g0_}{The intercept parameter of the detection function. Supplied when appropriate (for an independent linked hazard).}

\item{lambda0_}{The intercept parameter of the encounter rate function. Supplied when appropriate (for an independent unlinked 
or dependent hazard).}

\item{sigma_}{The range parameter of the detection / encounter rate function. Will be NULL when a dependent hazard is used.}

\item{endpoint_}{Specifies the duration of the simulation survey.}

\item{setup.hour_}{Used when the survey starts at some point within the first cycle.}

\item{occ.interval_}{The duration of an occasion. Used to construct the discrete-time secr capthist object.}

\item{detector_type_}{Specifies the type of detector. Defaults to using a "proximity" or passive detector but can also simulate
from an array of single-catch traps.}
}
\value{
Returns the capture histories with continuous times as a data frame, and with an occasion structure as a secr capthist object.
}
\description{
Simulates continuous-time datasets.
}
