% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{sim.gof}
\alias{sim.gof}
\title{Simulate sample of a GoF test statistic.}
\usage{
sim.gof(fit, statfn = Devdf, nsim = 99, newfit = FALSE, seed = NULL, ...)
}
\arguments{
\item{fit}{An \code{secr} fitted model. (This must have an element \code{$mask}.)}

\item{statfn}{A function that calculates a  GoF test statistic from a fitted model or 
\code{capthist} object.}

\item{nsim}{The number of simulations to do.}

\item{newfit}{If TRUE, refits model to every simulated dataset and recalculates expected numbers.}

\item{seed}{Random number seed (setting it gives repeatability of results).}

\item{...}{Other arguments to pass to statfn.}
}
\description{
Repeatedly simulates a population and capture history using the parameters of a 
fitted \code{secr} model as truth, fits a model of the same sort to the data,
and calculates the test statistic (specified via\code{statfn}) for each new
fitted model.

Only works with full likelhood models (those that include a model for density), and 
does not work for all of them. In particular, it works for models without
individual effects, except that it does work for a model with sex as a 
partially-observed individual effect in which both sexes have a common density.

Has only been tested with single-occasion surveys.
}
\examples{
require(secr)
sim.gof(secrdemo.0)

}
