% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{sim.time.gam}
\alias{sim.time.gam}
\title{Simulate capture time.}
\usage{
sim.time.gam(CDF.tot_, study.vec_, occ.interval_ = 24,
  haz.interval_ = 24, release.hour_ = 0)
}
\arguments{
\item{CDF.tot_}{The CDF that is used to simulate capture times via the probability integral transform approach.}

\item{study.vec_}{A vector of times that correspond to the time period one wants to simulate captures for.}

\item{occ.interval_}{The duration of an occasion.}

\item{haz.interval_}{The duration of a single hazard cycle.}

\item{release.hour_}{Relevant for SC traps when individuals are released at a particular time. Usually in such cases the 
release time will define the occasions.}
}
\value{
Returns both the actual capture time in total time units, and in terms of the hazard cycle. Also returns the 
capture occasion.
}
\description{
Called by the simctsCH function to simulate individual capture times using the CDF of capture times.
}
