% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrplotting.r
\name{splot.Dsurface}
\alias{splot.Dsurface}
\title{Plots density surface image}
\usage{
splot.Dsurface(fit, se.D = FALSE, cl.D = FALSE, alpha = 0.05, ...)
}
\arguments{
\item{fit}{is an object of class `secr'}

\item{se.D}{If TRUE, returns std error of estimated density}

\item{cl.D}{If TRUE, returns confidence bounds for density}

\item{alpha}{Significance level}

\item{...}{other arguments to be passed to \code{plot.SpatialPixelsDataFrame}}
}
\description{
Plots density surface image
}
