% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrplotting.r
\name{splotcovariate}
\alias{splotcovariate}
\title{Plots mask covariate using package sp functions}
\usage{
splotcovariate(mask, covariate, ...)
}
\arguments{
\item{mask}{is an object of class `mask'}

\item{covariate}{is a character variable with the name of one of the covariates in mask (the one to plot)}

\item{...}{other arguments to be passed to \code{sp} \code{plot} function.}
}
\description{
Plots mask covariate using package sp functions
}
