% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{streplace}
\alias{streplace}
\title{Replaces substring(s) in a string}
\usage{
streplace(string, from, to)
}
\arguments{
\item{string}{the string (character variable) in which replacing is to occur.}

\item{from}{a (vector of) character variables that are to be replaced.}

\item{to}{a (vector of) character variables of the same length as \code{from}
containing the strings that are to replace \code{from}.}
}
\description{
Replaces all instances of specified substrings in a string (character variable) 
 with specified other substrings.
}
