% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTfns.R
\name{study.haz.vec}
\alias{study.haz.vec}
\title{Construct the hazard vector for the full study.}
\usage{
study.haz.vec(cycle.vec_, y.vec.full_, spline.coeff_ = NULL,
  release.hour_, endpoint_, n.haz.cycles_, haz.interval_, gam.obj_,
  fit.obj_ = NULL)
}
\arguments{
\item{cycle.vec_}{A vector of times that correspond to a single hazard cycle.}

\item{y.vec.full_}{A vector of detection hazards that correspond to a single hazard cycle.}

\item{spline.coeff_}{A vector of spline coefficients.}

\item{release.hour_}{Primarily applicable to SC traps when the release event defines the occasions. Can also be used to 
adjust the first cycle if it starts after the cycle's start point.}

\item{endpoint_}{Specifies the duration of the simulation survey.}

\item{n.haz.cycles_}{The number of full hazard cycles in the full study period.}

\item{haz.interval_}{The duration of a single hazard cycle.}

\item{gam.obj_}{GAM obj used by function pred.Y to predict detection hazards.}

\item{fit.obj_}{A fitted glm object used by function pred.Y to predict detection hazards.}
}
\value{
Returns a vector of detection hazards of the same dimension as the supplied study time vector.
}
\description{
Constructs the hazard vector for the full study duration from the hazard for a single cycle. Used in both the LL
function (CT.SCR) and in the simulation function (sim.ctsCH).
}
