% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{wolverines}
\alias{wolverines}
\title{Wolverine baited camera trap dataset}
\usage{
data("wolverines")
}
\description{
Data are from Royle et al. (2011), from a camera trap study of wolverines in shouteastern Alaska in 
2008. Data are in a list that contains the following objects:
\describe{
    \item{\code{w.ch}:}{A capture history object (class
    \code{capthist}) containing the binary capture data for each
    of 21 individuals at each of 37 detectors on each of 165 occasions.
    The object \code{traps(w.ch)} is an object of class \code{traps}, that has
    covariates \code{elevation} and \code{stdelev}, containing 
    the elevation and the standardised elevation at each trap (standardised
    such that the elevations in \code{w.mesh} have mean zero and variance 1).}
    \item{\code{w.ch.count}:}{A capture history object (class
    \code{capthist}) containing the counts of captures of each
    of 21 individuals at each of 37 detectors across all 165 occasions.
    The object \code{traps(w.ch.count)} is an object of class \code{traps}, that has
    covariates \code{elevation} and \code{stdelev}, containing 
    the elevation and the standardised elevation at each trap (standardised
    such that the elevations in \code{w.mesh} have mean zero and variance 1).}
    \item{\code{w.mesh}:}{A capture history object (class
    \code{mask}) containing a mesh that extends 40km beyond the traps. 
    The mesh has covariates \code{elevation} and \code{stdelev}, containing 
    the elevation and the standardised elevation at each mesh point (standardised
    such that the elevations in \code{w.mesh} have mean zero and variance 1).}
}
}
\references{
Royle, J. A. and Kagoun, A. J. and Gardner, B. and Valkenburg, P. and Lowell, R. E. (2011) 
Density estimation in a wolverine population using spatial capture-recapture models. 
\textit{Journal of Wildlife Management} \textbf{75}: 604-611.
}
