% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{wolverinespatial}
\alias{wolverinespatial}
\title{Spatial objects to go with \code{wolerines} data}
\usage{
data("wolverinespatial")
}
\description{
Data are from Royle et al. (2011), from a camera trap study of wolverines in shouteastern Alaska in 
2008. Data are in a list that contains the following objects:
\describe{
    \item{\code{w.elev.spdf}:}{A \code{SpatialPixelsDataFrame}) containing the
    elevation data in the vicininty of the woleverine traps.}
    \item{\code{w.elev.spgf}:}{A \code{SpatialGridDataFrame}) containing the
    elevation data in the vicininty of the woleverine traps.}
    \item{\code{w.cl}:}{A \code{SpatialLinesDataFrame}) containing the
    Alaska coastline and land border in the vicininty of the woleverine traps.}
    \item{\code{w.traps.spdf}:}{A \code{SpatialPixelsDataFrame}) containing the 
    trap locations.}
}
}
