% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{plam}
\alias{plam}
\title{Calculate detection function or encounter rate parameters}
\usage{
plam(detectfn, pars, dists, pars.vcv = NULL, details = NULL,
  rgr = FALSE, hazard = FALSE)
}
\arguments{
\item{detectfn}{integer code or character string for shape of detection 
function 0 = halfnormal etc. – see \link{secr::detectfn}.}

\item{pars}{list, vector or matrix of parameter values.}

\item{dists}{distances at which to calculate the function.}

\item{pars.vcv}{variance-covariance matrix of pars}

\item{details}{list of ancillary parameters - see \link{secr::detectfn} for 
details}

\item{rgr}{logical; if TRUE a scaled curve r.g(r) is plotted instead of g(r)}

\item{hazard}{logical; if TRUE the hazard of detection (the expected encounter
function) is plotted instead of the probability.}
}
\description{
Calculates detection function or expected encounter rate at a range of
 distances, given detection function or encounter rate (hazard) function
 parameters. This is a very slightly hacked version of the \link{secr} function
 \link{secr::detectfnplot}, modified so as not to do the plotting and always 
 return the function values.
}
