% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrplotting.r
\name{plotER3d}
\alias{plotER3d}
\title{Plots encounter rate in 3D}
\usage{
plotER3d(capthist, mask = NULL, occasions = 1, add.points = TRUE,
  add.text = FALSE, add.maskedge = FALSE, add.mask = TRUE,
  add.traps = TRUE, maskcov = NULL, maskcol = parula, ncols = 40,
  ...)
}
\description{
Plots encounter rate in 3D
}
